requirements.txt
torch
sentence-transformers
faiss-cpu
# scann
datasets
rank-bm25
spacy
rapidfuzz
python-Levenshtein
numpy
pandas
matplotlib
tqdm
